+++ tests/integration_tests/cross_platform_shell_execution.rs
