+++ tests/integration_tests/installation_tests.rs
